glfw

+++ projects/aergia/src/io/graphics_display.cpp
@@ -74,7 +74,7 @@ namespace aergia {
   }
 
   ponos::Point2 GraphicsDisplay::getMouseNPos() {
-    float viewport[] = {0, 0, width, height};
+    int viewport[] = {0, 0, width, height};
     ponos::Point2 mp = getMousePos();
     return ponos::Point2((mp.x - viewport[0]) / viewport[2] * 2.0 - 1.0,
                          (mp.y - viewport[1]) / viewport[3] * 2.0 - 1.0);
